Update README.md
Processing temperatures with functions

Write a program that will input temperatures for consecutive days. It is similar to lab 24.4, except now the program will call a function that will return the average of the temperatures. It will also call a function that will return the highest temperature and a function that will return the lowest temperature. There will be no more than 50 temperatures.

First, the user will input the number of temperatures to be read. The program will store these values into an array and then average all the temperatures. It will also determine the highest temperature and the lowest temperature. The average should be displayed to two decimal places.

Sample Run:
Please input the number of temperatures to be read
5
Input temperature 1:
68
Input temperature 2:
75
Input temperature 3:
36
Input temperature 4:
91
Input temperature 5:
84
The average temperature is 70.80
The highest temperature is 91.00
The lowest temperature is 36.00